fix: cast heading level to long for compatibility with Math.clamp
- Ensures proper type handling during heading level adjustment in `HtmlProcessor`.

+++ src/main/java/ch/sbb/polarion/extension/docx_exporter/util/HtmlProcessor.java
@@ -237,7 +237,7 @@ public class HtmlProcessor {
                 heading.addClass("title");
             } else {
                 int level = heading.tagName().charAt(1) - '0';
-                int newLevel = Math.clamp(level - 1, 1, 6);
+                int newLevel = Math.clamp((long) level - 1, 1, 6);
                 heading.tagName("h" + newLevel);
             }
         }
